+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Romantic Versioning](http://blog.legacyteam.info/20
 
 ## [Unreleased]
 
+## [0.9.18] - 2021/05/08
+
+### Added
+- unload feature for URLCaching.
+
+### Changed
+- enable to select Destroy cached Object when unload/purge on URLCaching feature.
+
+
+
 ## [0.9.17] - 2021/04/26
 
 ### Changed
@@ -333,7 +343,8 @@ and this project adheres to [Romantic Versioning](http://blog.legacyteam.info/20
 - Something.
 
 
-[Unreleased]: https://github.com/sassembla/autoya/compare/0.9.17...HEAD
+[Unreleased]: https://github.com/sassembla/autoya/compare/0.9.18...HEAD
+[0.9.18]: https://github.com/sassembla/autoya/compare/0.9.17...0.9.18
 [0.9.17]: https://github.com/sassembla/autoya/compare/0.9.16...0.9.17
 [0.9.16]: https://github.com/sassembla/autoya/compare/0.9.15...0.9.16
 [0.9.15]: https://github.com/sassembla/autoya/compare/0.9.14...0.9.15